Split create from the application instance views
