+++ lms/views/admin/application_instance/view.py
@@ -17,27 +17,6 @@ from lms.views.admin.application_instance._core import (
 )
 
 
-class NewAppInstanceSchema(PyramidRequestSchema):
-    """Schema for creating a new application instance."""
-
-    location = "form"
-
-    developer_key = fields.Str(required=False, allow_none=True)
-    developer_secret = fields.Str(required=False, allow_none=True)
-
-    name = fields.Str(required=True, validate=validate.Length(min=1))
-    lms_url = fields.URL(required=True)
-    email = fields.Email(required=True)
-    organization_public_id = fields.Str(required=True, validate=validate.Length(min=1))
-
-
-class NewAppInstanceSchemaV13(NewAppInstanceSchema):
-    """Schema for creating a new LTI 1.3 application instance."""
-
-    deployment_id = fields.Str(required=True, validate=validate.Length(min=1))
-    lti_registration_id = fields.Str(required=True)
-
-
 class UpdateApplicationInstanceSchema(PyramidRequestSchema):
     """Schema for updating an application instance."""
 
@@ -65,66 +44,6 @@ class AdminApplicationInstanceViews(BaseApplicationInstanceView):
             LTIRegistrationService
         )
         self._aes_service = request.find_service(AESService)
-
-    @view_config(
-        route_name="admin.instance.new",
-        renderer="lms:templates/admin/application_instance/new.html.jinja2",
-    )
-    def new_instance_start(self):
-        """Show the page to kick off creating a new application instance."""
-
-        lti_registration = None
-        if lti_registration_id := self.request.params.get("lti_registration_id"):
-            lti_registration = self.lti_registration_service.get_by_id(
-                lti_registration_id.strip()
-            )
-
-        return dict(self.request.params, lti_registration=lti_registration)
-
-    @view_config(route_name="admin.instance.new", request_method="POST")
-    def new_instance_callback(self):
-        """Create an application instance (callback from the new AI page)."""
-
-        lti_registration_id = self.request.params.get("lti_registration_id", "").strip()
-        lti_registration_id = int(lti_registration_id) if lti_registration_id else None
-
-        if flash_validation(
-            self.request,
-            NewAppInstanceSchemaV13 if lti_registration_id else NewAppInstanceSchema,
-        ):
-            # Looks like something went wrong!
-            return self._redirect("admin.instance.new", _query=self.request.params)
-
-        try:
-            ai = self.application_instance_service.create_application_instance(
-                name=self.request.params["name"].strip(),
-                lms_url=self.request.params["lms_url"].strip(),
-                email=self.request.params["email"].strip(),
-                deployment_id=self.request.params.get("deployment_id", "").strip(),
-                developer_key=self.request.params.get("developer_key", "").strip(),
-                developer_secret=self.request.params.get(
-                    "developer_secret", ""
-                ).strip(),
-                organization_public_id=self.request.params.get(
-                    "organization_public_id", ""
-                ).strip(),
-                lti_registration_id=lti_registration_id,
-            )
-        except InvalidPublicId as err:
-            self.request.session.flash(
-                {"organization_public_id": [str(err)]}, "validation"
-            )
-
-            return self._redirect("admin.instance.new", _query=self.request.params)
-        except IntegrityError:
-            self.request.session.flash(
-                f"Application instance with deployment_id: {self.request.params['deployment_id']} already exists",
-                "errors",
-            )
-
-            return self._redirect("admin.instance.new", _query=self.request.params)
-
-        return self._redirect("admin.instance", id_=ai.id)
 
     @view_config(
         route_name="admin.instance.upgrade",
